+++ tests/testthat/test-utils.R
@@ -123,10 +123,13 @@ test_that("fit0 doesn't alter anything", {
   expect_identical(model, fit(model, 0, displayProgress = FALSE))
 })
 
+scores <- compute_PicaultRenault_scores(ECB_press_conferences)
+EC <- scores[names(ECB_press_conferences_tokens), "EC"]
 
 test_that("merge_topics works", {
   toks <- ECB_press_conferences_tokens[1:50]
   model <- LDA(toks)
+  sentopics_sentiment(model) <- EC[names(model$tokens)]
   merged <- merge_topics(model, as.list(1:5))
   sentopics_labels(merged)
   sentopics_labels(merged) <- NULL