+++ NEWS.md
@@ -21,11 +21,17 @@
 * Modified names of columns in the `PicaultRenault_data` dataset to
   syntactically valid names.
 
+* Removed `.sentiment` docvars in `ECB_press_conferences` dataset.
+
 ### Other changes
 
 * Reformatted codebase using Air.
 * Removed a comprehensive assessment from the `PicaultRenault_data` dataset
   not present in the original paper.
+
+# sentopics 0.7.5
+
+* Fixed an example following change in suggested dependency.
 
 # sentopics 0.7.5
 
@@ -39,7 +45,7 @@
 
 * Renamed `grow()` to `fit()` as a more intuitive name. `grow()` remains in the package for compatibility with older version.
 * Aligned the weight argument and the algorithm of `topWords(..., method="FREX")` to the original paper.
-* Fixed the un-exported function `get_ECB_conferences()` to accommodates changes from the ECB website.
+* Fixed the unexported function `get_ECB_conferences()` to accommodates changes from the ECB website.
 * Fixed the model conversion from newer `seededlda` versions (1.2.0 and subsequent)
 
 # sentopics 0.7.2